+++ brownie/project/main.py
@@ -1,6 +1,7 @@
 #!/usr/bin/python3
 
 import json
+import os
 import shutil
 import sys
 import zipfile
@@ -17,13 +18,22 @@ from brownie._config import (
     _load_project_config,
 )
 from brownie.exceptions import ProjectAlreadyLoaded, ProjectNotFound
-from brownie.network.contract import ContractContainer
+from brownie.network.contract import ContractContainer, ProjectContract
+from brownie.network.state import _remove_contract
 from brownie.project import compiler
 from brownie.project.build import BUILD_KEYS, Build
 from brownie.project.sources import Sources, get_hash
 from brownie.utils import color
 
-FOLDERS = ["contracts", "scripts", "reports", "tests", "build", "build/contracts"]
+FOLDERS = [
+    "contracts",
+    "scripts",
+    "reports",
+    "tests",
+    "build",
+    "build/contracts",
+    "build/deployments",
+]
 MIXES_URL = "https://github.com/brownie-mix/{}-mix/archive/master.zip"
 
 _loaded_projects = []
@@ -138,6 +148,7 @@ class Project(_ProjectBase):
         self._compiler_config["version"] = solc_version
         self._compile(changed, self._compiler_config, False)
         self._create_containers()
+        self._load_deployments()
 
         # add project to namespaces, apply import blackmagic
         name = self._name
@@ -176,6 +187,26 @@ class Project(_ProjectBase):
             (True for k, v in build_json["compiler"].items() if config[k] and v != config[k]), False
         )
 
+    def _load_deployments(self) -> None:
+        if not CONFIG["active_network"].get("persist", None) or self._project_path is None:
+            return
+        network = CONFIG["active_network"]["name"]
+        path = self._project_path.joinpath(f"build/deployments/{network}")
+        path.mkdir(exist_ok=True)
+        deployments = list(
+            self._project_path.glob(f"build/deployments/{CONFIG['active_network']['name']}/*.json")
+        )
+        deployments.sort(key=os.path.getmtime)
+        for build_json in deployments:
+            with build_json.open() as fp:
+                build = json.load(fp)
+            if build["contractName"] not in self._containers:
+                build_json.unlink()
+            else:
+                container = self._containers[build["contractName"]]
+                contract = ProjectContract(self, build, build_json.stem)
+                container._contracts.append(contract)
+
     def _update_and_register(self, dict_: Any) -> None:
         dict_.update(self)
         self._namespaces.append(dict_)
@@ -203,6 +234,13 @@ class Project(_ProjectBase):
                 if v == self or (k in self and v == self[k])  # type: ignore
             ]:
                 del dict_[key]
+
+        # remove contracts
+        for contract in [x for v in self._containers.values() for x in v._contracts]:
+            _remove_contract(contract)
+        for container in self._containers.values():
+            container._contracts.clear()
+        self._containers.clear()
 
         # undo black-magic
         name = self._name